+++ api/instances/bundled-workbench-adapters.ts
@@ -1,6 +1,3 @@
-import { localWorkbenchAdapterRegistration } from './garbage/workbench.js'
 import type { LocalWorkbenchAdapterRegistration } from '../workbench/local-adapter-types'
 
-export const bundledLocalWorkbenchAdapterRegistrations: LocalWorkbenchAdapterRegistration[] = [
-  localWorkbenchAdapterRegistration,
-]
+export const bundledLocalWorkbenchAdapterRegistrations: LocalWorkbenchAdapterRegistration[] = []