+++ api/instances/garbage/newsletter-store.ts
@@ -1,1 +1,0 @@
-export * from '../../../../apps/garbage/src/api/newsletter-store.ts'