+++ api/instances/garbage/newsletter/draft.ts
@@ -1,1 +1,0 @@
-export { default } from '../../../../../apps/garbage/src/api/newsletter/draft.ts'