+++ api/instances/garbage/workbench.ts
@@ -1,1 +1,0 @@
-export * from '../../../../apps/garbage/src/api/workbench.ts'